adds EmailReadImapV2 with fixes
