+++ packages/nodes-base/credentials/Imap.credentials.ts
@@ -38,5 +38,32 @@ export class Imap implements ICredentialType {
 			type: 'boolean',
 			default: true,
 		},
+		{
+			displayName: 'Allow Self-Signed Certificates',
+			name: 'allowUnauthorizedCerts',
+			type: 'boolean',
+			description: 'Whether to connect even if SSL certificate validation is not possible',
+			default: false,
+		},
 	];
+}
+
+export interface ICredentialsDataImap {
+	host: string;
+	port: number;
+	user: string;
+	password: string;
+	secure: boolean;
+	allowUnauthorizedCerts: boolean;
+}
+
+export function isCredentialsDataImap(candidate: unknown): candidate is ICredentialsDataImap {
+	const o = candidate as ICredentialsDataImap;
+	return (
+		o.host !== undefined &&
+		o.password !== undefined &&
+		o.port !== undefined &&
+		o.secure !== undefined &&
+		o.user !== undefined
+	);
 }
